[MOV] im_livechat: move abstractThread out of public_livechat.js
Extract AbstractThread from public_livechat.js, and move it
to public_livechat_thread.js.
The goal is to split the public_livechat.js file into multiple
parts, to prepare for refactoring.
